--- client/src/components/pages/AccountPage/utils.js
@@ -1,24 +1,0 @@
-export function gatherCurrencySymbol(currencyCode) {
-  switch (currencyCode) {
-    case 'USD':
-      return '$';
-    case 'GBP':
-      return '£';
-    default:
-      return '*';
-  }
-}
-
-export function percIncrease(a, b) {
-  let percent;
-  if (b !== 0) {
-    if (a !== 0) {
-      percent = ((b - a) / a) * 100;
-    } else {
-      percent = b * 100;
-    }
-  } else {
-    percent = -a * 100;
-  }
-  return percent.toFixed(2);
-}
